rename files and reorganize content
