+++ dockertests/utils/__init__.py
@@ -6,10 +6,10 @@ Utilities for Azure Functions container testing.
 
 from .functions_container_controller import FunctionsContainerController
 from .azurite_container_controller import AzuriteContainerController
-from .functions_test_environment import FunctionsTestEnvironment
+from .linux_consumption_test_environment import LinuxConsumptionTestEnvironment
 
 __all__ = [
     'FunctionsContainerController',
     'AzuriteContainerController',
-    'FunctionsTestEnvironment',
+    'LinuxConsumptionTestEnvironment',
 ]
